backend/requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
filelock==3.0.12
future==0.18.2
h11==0.11.0
idna==2.10
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.790
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.2
packaging==20.4
Pillow==7.2.0
psycopg2-binary==2.8.6
pydantic==1.6.1
pylint==2.6.0
pyparsing==2.4.7
regex==2020.10.11
requests==2.24.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
sentence-transformers==0.3.7.2
sentencepiece==0.1.91
six==1.15.0
starlette==0.13.6
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
transformers==3.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.12.1
wrapt==1.12.1
